tests(e2e): synced with latest liveblog changes

+++ spec/setup.js
@@ -9,10 +9,13 @@ var resetApp = require('./helpers/fixtures').resetApp;
 beforeEach(function(done) {
     getToken(function() {
         resetApp(function() {
-            browser.get('/');
-            browser.executeScript('sessionStorage.clear();localStorage.clear();');
-            protractor.getInstance().waitForAngular();
-            done();
+            browser.driver.getCurrentUrl().then(function(url) {
+                if (url.indexOf('data:') !== 0) {
+                    browser.executeScript('sessionStorage.clear();localStorage.clear();');
+                    protractor.getInstance().waitForAngular();
+                }
+                done();
+            });
         });
     });
 });
